+++ Package.swift
@@ -4,10 +4,11 @@ import PackageDescription
 let package = Package(
     name: "SwiftMessages",
     platforms: [
-        .iOS("9.0")
+        .iOS("13.0")
     ],
     products: [
-        .library(name: "SwiftMessages", targets: ["SwiftMessages"])
+        .library(name: "SwiftMessages", targets: ["SwiftMessages"]),
+        .library(name: "SwiftMessages-Dynamic", type: .dynamic, targets: ["SwiftMessages"])
     ],
     targets: [
         .target(