+++ SwiftMessages/Animator.swift
@@ -58,6 +58,7 @@ public class AnimationContext {
     }
 }
 
+@MainActor
 public protocol Animator: AnyObject {
 
     /// Adopting classes should declare as `weak`.