+++ SwiftMessages/BaseView.swift
@@ -280,6 +280,11 @@ open class BaseView: UIView, BackgroundViewable, MarginAdjustable {
 
     private var layoutConstraints: [NSLayoutConstraint] = []
     private var regularWidthLayoutConstraints: [NSLayoutConstraint] = []
+
+    open override func layoutSubviews() {
+        super.layoutSubviews()
+        updateShadowPath()
+    }
 }
 
 /*
@@ -293,7 +298,7 @@ extension BaseView {
     /// because the background view may be masked. So, when modifying the drop shadow,
     /// be sure to set the shadow properties of this view's layer. The shadow path is
     /// updated for you automatically.
-    open func configureDropShadow() {
+    public func configureDropShadow() {
         layer.shadowColor = UIColor.black.cgColor
         layer.shadowOffset = CGSize(width: 0.0, height: 2.0)
         layer.shadowRadius = 6.0
@@ -303,7 +308,7 @@ extension BaseView {
     }
 
     /// A convenience function to turn off drop shadow
-    open func configureNoDropShadow() {
+    public func configureNoDropShadow() {
         layer.shadowOpacity = 0
     }
 
@@ -333,11 +338,6 @@ extension BaseView {
         } else {
             // Update the layer's `shadowPath` without animation
             layer.shadowPath = shadowPath        }
-    }
-
-    open override func layoutSubviews() {
-        super.layoutSubviews()
-        updateShadowPath()
     }
 }