+++ SwiftMessages/MarginAdjustable+Extensions.swift
@@ -13,25 +13,8 @@ extension MarginAdjustable where Self: UIView {
         var layoutMargins: UIEdgeInsets = layoutMarginAdditions
         var safeAreaInsets: UIEdgeInsets = {
             guard respectSafeArea else { return .zero }
-            if #available(iOS 11, *) {
-                insetsLayoutMarginsFromSafeArea = false
-                return self.safeAreaInsets
-            } else {
-                #if SWIFTMESSAGES_APP_EXTENSIONS
-                let application: UIApplication? = nil
-                #else
-                let application: UIApplication? = UIApplication.shared
-                #endif
-                if !context.safeZoneConflicts.isDisjoint(with: [.statusBar]),
-                   let app = application,
-                   app.statusBarOrientation == .portrait || app.statusBarOrientation == .portraitUpsideDown {
-                    let frameInWindow = convert(bounds, to: window)
-                    let top = max(0, 20 - frameInWindow.minY)
-                    return UIEdgeInsets(top: top, left: 0, bottom: 0, right: 0)
-                } else {
-                    return .zero
-                }
-            }
+            insetsLayoutMarginsFromSafeArea = false
+            return self.safeAreaInsets
         }()
         if !context.safeZoneConflicts.isDisjoint(with: .overStatusBar) {
             safeAreaInsets.top = 0