+++ SwiftMessages/PhysicsAnimation.swift
@@ -8,6 +8,7 @@
 
 import UIKit
 
+@MainActor
 public class PhysicsAnimation: NSObject, Animator {
 
     public enum Placement {
@@ -93,9 +94,7 @@ public class PhysicsAnimation: NSObject, Animator {
         guard let adjustable = messageView as? MarginAdjustable & UIView,
             let context = context else { return }
         adjustable.preservesSuperviewLayoutMargins = false
-        if #available(iOS 11, *) {
-            adjustable.insetsLayoutMarginsFromSafeArea = false
-        }
+        adjustable.insetsLayoutMarginsFromSafeArea = false
         adjustable.layoutMargins = adjustable.defaultMarginAdjustment(context: context)
     }