+++ SwiftMessages/Presenter.swift
@@ -12,6 +12,7 @@ protocol PresenterDelegate: AnimationDelegate {
     func hide(presenter: Presenter)
 }
 
+@MainActor
 class Presenter: NSObject {
 
     // MARK: - API
@@ -68,8 +69,16 @@ class Presenter: NSObject {
         return duration
     }
 
+    /// Detects the scenario where the view was shown, but the containing view heirarchy was removed before the view
+    /// was hidden. This unusual scenario could result in the message queue being blocked because the presented
+    /// view was not properly hidden by SwiftMessages. `isOrphaned` allows the queuing logic to unblock the queue.
+    var isOrphaned: Bool {
+        return installed && view.window == nil
+    }
+
     // MARK: - Constants
 
+    @MainActor
     enum PresentationContext {
         case viewController(_: Weak<UIViewController>)
         case view(_: Weak<UIView>)
@@ -97,7 +106,7 @@ class Presenter: NSObject {
 
     private weak var delegate: PresenterDelegate?
     private var presentationContext = PresentationContext.viewController(Weak<UIViewController>(value: nil))
-
+    private var installed = false
     private var interactivelyHidden = false;
 
     // MARK: - Showing and hiding
@@ -232,7 +241,7 @@ class Presenter: NSObject {
     }
 
     private func safeZoneConflicts() -> SafeZoneConflicts {
-        guard let window = maskingView.window else { return [] }
+        guard let _ = maskingView.window else { return [] }
         let windowLevel: UIWindow.Level = {
             if let vc = presentationContext.viewControllerValue() as? WindowViewController {
                 return vc.config.windowLevel ?? .normal
@@ -249,47 +258,34 @@ class Presenter: NSObject {
             if let vc = presentationContext.viewControllerValue() as? UITabBarController { return vc.sm_isVisible(view: vc.tabBar) }
             return false
         }()
-        if #available(iOS 11, *) {
-            if windowLevel > .normal {
-                // TODO seeing `maskingView.safeAreaInsets.top` value of 20 on
-                // iPhone 8 with status bar window level. This seems like an iOS bug since
-                // the message view's window is above the status bar. Applying a special rule
-                // to allow the animator to revove this amount from the layout margins if needed.
-                // This may need to be reworked if any future device has a legitimate 20pt top safe area,
-                // such as with a potentially smaller notch.
-                if maskingView.safeAreaInsets.top == 20 {
-                    return [.overStatusBar]
-                } else {
-                    var conflicts: SafeZoneConflicts = []
-                    if maskingView.safeAreaInsets.top > 0 {
-                        conflicts.formUnion(.sensorNotch)
-                    }
-                    if maskingView.safeAreaInsets.bottom > 0 {
-                        conflicts.formUnion(.homeIndicator)
-                    }
-                    return conflicts
+        if windowLevel > .normal {
+            // TODO seeing `maskingView.safeAreaInsets.top` value of 20 on
+            // iPhone 8 with status bar window level. This seems like an iOS bug since
+            // the message view's window is above the status bar. Applying a special rule
+            // to allow the animator to revove this amount from the layout margins if needed.
+            // This may need to be reworked if any future device has a legitimate 20pt top safe area,
+            // such as with a potentially smaller notch.
+            if maskingView.safeAreaInsets.top == 20 {
+                return [.overStatusBar]
+            } else {
+                var conflicts: SafeZoneConflicts = []
+                if maskingView.safeAreaInsets.top > 0 {
+                    conflicts.formUnion(.sensorNotch)
                 }
+                if maskingView.safeAreaInsets.bottom > 0 {
+                    conflicts.formUnion(.homeIndicator)
+                }
+                return conflicts
             }
-            var conflicts: SafeZoneConflicts = []
-            if !underNavigationBar {
-                conflicts.formUnion(.sensorNotch)
-            }
-            if !underTabBar {
-                conflicts.formUnion(.homeIndicator)
-            }
-            return conflicts
-        } else {
-            #if SWIFTMESSAGES_APP_EXTENSIONS
-            return []
-            #else
-            if UIApplication.shared.isStatusBarHidden { return [] }
-            if (windowLevel > UIWindow.Level.normal) || underNavigationBar { return [] }
-            let statusBarFrame = UIApplication.shared.statusBarFrame
-            let statusBarWindowFrame = window.convert(statusBarFrame, from: nil)
-            let statusBarViewFrame = maskingView.convert(statusBarWindowFrame, from: nil)
-            return statusBarViewFrame.intersects(maskingView.bounds) ? SafeZoneConflicts.statusBar : []
-            #endif
         }
+        var conflicts: SafeZoneConflicts = []
+        if !underNavigationBar {
+            conflicts.formUnion(.sensorNotch)
+        }
+        if !underTabBar {
+            conflicts.formUnion(.homeIndicator)
+        }
+        return conflicts
     }
 
     private func getPresentationContext() throws -> PresentationContext {
@@ -415,6 +411,7 @@ class Presenter: NSObject {
             maskingView.accessibleElements = elements
         }
 
+        installed = true
         guard let containerView = presentationContext.viewValue() else { return }
         (presentationContext.viewControllerValue() as? WindowViewController)?.install()
         installMaskingView(containerView: containerView)